Añadi mejoras para eliminar y protegi rutas para solo mostrar si esta logeado

+++ src/main/java/org/rivera/apiservlet/webapp/carrito/controllers/EliminarProductoServlet.java
@@ -5,11 +5,13 @@ import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import org.rivera.apiservlet.webapp.carrito.models.Producto;
 import org.rivera.apiservlet.webapp.carrito.service.ProductoService;
 import org.rivera.apiservlet.webapp.carrito.service.ProductoServiceJdbcImp;
 
 import java.io.IOException;
 import java.sql.Connection;
+import java.util.Optional;
 
 @WebServlet("/productos/eliminar")
 public class EliminarProductoServlet extends HttpServlet {
@@ -25,8 +27,16 @@ public class EliminarProductoServlet extends HttpServlet {
       id = 0L;
     }
     if( id > 0 ) {
-      service.deleteProduct(id);
+      //La parte de encontrar por "ID" es para mejorar el proceso y estar seguro de la existencia del producto en BD
+      Optional<Producto> op = service.byIdProduct(id);
+      if( op.isPresent() ) {
+        service.deleteProduct(id);
+        resp.sendRedirect(req.getContextPath() + "/productosSession.html");
+      } else {
+        resp.sendError(HttpServletResponse.SC_NOT_FOUND, "No existe ID del producto");
+      }
+    } else {
+      resp.sendError(HttpServletResponse.SC_NOT_FOUND, "El ID debe enviarse en la ruta URL");
     }
-    resp.sendRedirect(req.getContextPath() + "/productosSession.html");
   }
 }
